remove "r"::Symbol from type
The translation in src/interp/ax.boot can currently not cope with
  UnivariatePolynomial("r"::Sy, FI)
It will be wrongly translated into something involving "Enumeration".
The compilation for libaldor.al then aborts with

"ap/CKP.ap", line 28:
         (|Apply| |::| (|Apply| |Enumeration| |r|) |Symbol|)
..................^
[L28 C19] #1 (Error) There are no suitable meanings for the operator `::'.

The translation into an Enumeration type is clearly wrong here.

Removing the need for an explicit name of the polynomial variable seems to
be appropriate even without the above bug in ax.boot.

git-svn-id: https://svn.code.sf.net/p/fricas/code@1267 b0c55286-4f34-0410-a049-a1e7e93b0762